refactor(chaos): Unify effect API on apply()
Rename ChaosEffects to ChaosEffectsConfig so the TypedDict is no longer one letter from the ChaosEffect ABC. No deprecation alias and the package is pre-release.

Export ModelEffect and ModelEffectUnion so the model side of the effect hierarchy mirrors the tool side.

Fold cancel_message() into apply() on EmptyResponse and FullRefusal and drop the PreModelEffect Protocol and cast() it required. The overrides keep the ignored content param for LSP compatibility and narrow the return to str. EmptyResponse returns " " not "" because the SDK checks event.cancel for truthiness.

Tests for pre-effect apply() live in test_effects.py, ChaosCase validation in test_case.py, model hooks in test_plugin.py alongside the tool hooks, sharing its fixtures. No existing assertion changed.
